+++ src/main/java/com/gregtechceu/gtceu/integration/jade/GTJadePlugin.java
@@ -3,7 +3,7 @@ package com.gregtechceu.gtceu.integration.jade;
 import com.gregtechceu.gtceu.GTCEu;
 import com.gregtechceu.gtceu.api.blockentity.MetaMachineBlockEntity;
 import com.gregtechceu.gtceu.common.blockentity.FluidPipeBlockEntity;
-import com.gregtechceu.gtceu.common.data.GTItems;
+import com.gregtechceu.gtceu.common.data.GTMaterialItems;
 import com.gregtechceu.gtceu.integration.jade.provider.*;
 
 import net.minecraft.world.item.Item;
@@ -82,7 +82,7 @@ public class GTJadePlugin implements IWailaPlugin {
     }
 
     static {
-        GTItems.TOOL_ITEMS.columnMap().forEach((type, map) -> {
+        GTMaterialItems.TOOL_ITEMS.columnMap().forEach((type, map) -> {
             if (type.harvestTags.isEmpty() || type.harvestTags.get(0).location().getNamespace().equals("minecraft"))
                 return;
             HarvestToolProvider.registerHandler(new SimpleToolHandler(type.name, type.harvestTags.get(0),
